pyproject.toml
[project]
name = "line-extraction-deps-env"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "debugpy==1.8.14",
    "gitpython==3.1.43",
    "jinja2==3.1.6",
    "lxml==4.9.4",
    "matplotlib>=3.9.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "pybind11-stubgen==2.5.4",
    "pynput==1.8.1",
    "pyserial==3.5",
    "pytest>=8.0.0",
    "pyyaml==6.0.2",
    "requests>=2.32.4",
    "urllib3==1.26.6",
    "xmljson==0.2.1",
    "pre-commit>=3.0.0",
    "ruff>=0.4.0",
    "yamllint>=1.35.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "ipykernel>=6.29.0",
    "ipywidgets>=8.1.0",
    "ipympl>=0.9.0",
    "scikit-image>=0.21.0",
    # Rerun.io visualization
    "rerun-sdk[notebook]>=0.29.0",
    # PyTorch ecosystem for deep learning demos
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    # Segmentation models
    "segment-anything>=1.0",
    "ultralytics>=8.0.0",
]
